+++ src/main/java/com/alkemy/ong/config/controller/CategoryController.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.controller;
+package com.alkemy.ong.config.controller;
 
 import com.alkemy.ong.dto.CategoryDto;
 import com.alkemy.ong.model.entity.Category;
@@ -22,12 +22,9 @@ public class CategoryController {
     @Autowired
     private CategoryServiceImpl categoryService;
 
-    @Autowired
-    private ICategoryService iCategoryService;
-
     @PostMapping("/categories")
     public ResponseEntity<CategoryDto> save(@RequestBody CategoryDto categoryDto) {
-        CategoryDto newCategory = iCategoryService.save(categoryDto);
+        CategoryDto newCategory = CategoryService.save(categoryDto);
         return new ResponseEntity<>(newCategory, HttpStatus.OK);
     }
