+++ src/main/java/com/alkemy/ong/config/controller/CommentController.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.controller;
+package com.alkemy.ong.config.controller;
 
 import com.alkemy.ong.exception.OperationNotAllowedException;
 import com.alkemy.ong.service.abstraction.ICommentsService;
