+++ src/main/java/com/alkemy/ong/config/controller/NewsController.java
@@ -1,18 +1,15 @@
-package com.alkemy.ong.controller;
+package com.alkemy.ong.config.controller;
 
 import com.alkemy.ong.exception.FieldInvalidException;
+import com.fasterxml.jackson.databind.introspect.TypeResolutionContext.Empty;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.*;
 import com.alkemy.ong.dto.NewsDto;
 import com.alkemy.ong.service.abstraction.INewsService;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
+import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 
 @RestController
@@ -31,5 +28,11 @@ public class NewsController {
     @GetMapping("/{id}")
     public ResponseEntity<NewsDto> getNewsById(@PathVariable Long id) {
         return ResponseEntity.ok(newsService.findNewsById(id));
+    }
+
+    @DeleteMapping(value = "/{id}")
+    public ResponseEntity<Empty> delete(@PathVariable long id) throws EntityNotFoundException {
+        newsService.delete(id);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
