+++ src/main/java/com/alkemy/ong/service/NewsServiceImpl.java
@@ -1,6 +1,7 @@
 package com.alkemy.ong.service;
 
 import com.alkemy.ong.config.modelmapper.NewsMapper;
+import com.alkemy.ong.model.entity.Category;
 import com.alkemy.ong.model.entity.News;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -8,6 +9,7 @@ import com.alkemy.ong.repository.INewsRepository;
 import com.alkemy.ong.dto.NewsDto;
 import com.alkemy.ong.service.abstraction.INewsService;
 import javax.persistence.EntityNotFoundException;
+import java.util.Optional;
 
 @Service
 public class NewsServiceImpl implements INewsService {
@@ -33,5 +35,20 @@ public class NewsServiceImpl implements INewsService {
         News saved = newsRepository.save(newsEntity);
         NewsDto result = newsMapper.newsEntityToDto(saved);
         return result;
+    }
+
+    @Override
+    public void delete(Long id) throws EntityNotFoundException {
+        News news = getNews(id);
+        news.setSoftDelete(true);
+        newsRepository.save(news);
+    }
+
+    private News getNews(Long id) {
+    Optional<News> newsOptional = newsRepository.findById(id);
+    if (newsOptional.isEmpty() || newsOptional.get().isSoftDelete()) {
+        throw new EntityNotFoundException(NEWS_NOT_FOUND_MESSAGE);
+    }
+    return newsOptional.get();
     }
 }
