+++ src/main/java/com/alkemy/ong/service/abstraction/INewsService.java
@@ -3,6 +3,10 @@ package com.alkemy.ong.service.abstraction;
 import com.alkemy.ong.exception.FieldInvalidException;
 import com.alkemy.ong.exception.OperationNotAllowedException;
 import com.alkemy.ong.dto.NewsDto;
+import com.alkemy.ong.model.entity.News;
+
+import javax.persistence.EntityNotFoundException;
+import java.lang.reflect.Field;
 
 public interface INewsService {
 
@@ -10,4 +14,6 @@ public interface INewsService {
 
     NewsDto postNews(NewsDto newsDto) throws FieldInvalidException;
 
+    void delete(Long id) throws EntityNotFoundException;
+    ;
 }
